requirements.txt
Telethon==1.8.0
python-telegram-bot==11.1.0
loguru==0.3.2